requirements.txt
google-cloud-texttospeech
gtts
flask
uvicorn
jinja2
python-dotenv
sounddevice
pyaudio
numpy
pytz
schedule
websocket-client
google-generativeai
ephem
faster-whisper
openai
gTTS
edge-tts
pydub==0.25.1
requests
google-api-python-client
websocket-client
faster-whisper
asyncio

